Corrected issue about images and authentication

+++ backend/controllers/glasses.controller.js
@@ -79,15 +79,26 @@ exports.update = async (req, res) => {
       });
     }
 
-    if (existingGlasses.image) {
+    // Eliminar la imagen antigua solo si existe y se proporciona una nueva imagen
+    if (existingGlasses.image && req.file) {
       const imagePath = path.join(__dirname, '../public/images', existingGlasses.image);
-      fs.unlinkSync(imagePath);
+      if (fs.existsSync(imagePath)) {
+        fs.unlinkSync(imagePath);
+      } else {
+        console.warn("Image file does not exist:", imagePath);
+      }
     }
 
     if (req.file) {
-      const img = req.file.path;
+      const img = req.file.filename;
       glasses.image = img;
+
+      console.log("New image path:", path.join(__dirname, '../public/images', img));
+    } else {
+      glasses.image = existingGlasses.image;
     }
+
+    console.log("Existing image path:", path.join(__dirname, '../public/images', existingGlasses.image));
 
     await Glasses.update(glasses, {
       where: { id: id }
@@ -95,33 +106,45 @@ exports.update = async (req, res) => {
 
     res.status(200).send("Glasses updated");
   } catch (err) {
+    console.error(err);
     res.status(500).send({
-      message: "Error to update the glasses"
+      message: "Error updating the glasses"
     });
   }
 };
 
+
 exports.delete = async (req, res) => {
   try {
-      let id = req.params.id;
-      const existingGlasses = await Glasses.findByPk(id);
+    let id = req.params.id;
+    const existingGlasses = await Glasses.findByPk(id);
 
-      if (!existingGlasses) {
-          return res.status(404).send({
-              message: `Cannot find Glasses with id=${id}.`
-          });
-      }
-
-      if (existingGlasses.image) {
-          const imagePath = path.join(__dirname, '../public/images', existingGlasses.image);
-          fs.unlinkSync(imagePath);
-      }
-
-      await Glasses.destroy({ where: { id: id } });
-      res.status(200).json("Glasses has been deleted.");
-  } catch (err) {
-      res.status(500).send({
-          message: "Error deleting the glasses"
+    if (!existingGlasses) {
+      return res.status(404).send({
+        message: `Cannot find Glasses with id=${id}.`
       });
+    }
+
+    try {
+      if (existingGlasses.image) {
+        const imagePath = path.join(__dirname, '../public/images', existingGlasses.image);
+        if (fs.existsSync(imagePath)) {
+          fs.unlinkSync(imagePath);
+          console.log("Image file deleted:", imagePath);
+        } else {
+          console.warn("Image file does not exist:", imagePath);
+        }
+      }
+    } catch (err) {
+      console.error("Error deleting image file:", err);
+    }
+
+    await Glasses.destroy({ where: { id: id } });
+    res.status(200).json("Glasses has been deleted.");
+  } catch (err) {
+    console.error(err);
+    res.status(500).send({
+      message: "Error deleting the glasses"
+    });
   }
 };
